cit-api/requirements.txt
Django==2.2.16
psycopg2-binary==2.8.6
djangorestframework==3.12.1
djangorestframework-csv==2.1.0
djangorestframework-gis==0.16
django-cors-headers==3.5.0
drf-yasg==1.20.0
gunicorn==20.0.4
pyyaml==5.4
uritemplate==3.0.1
requests==2.24.0
jsonschema==3.2.0
msal==1.5.1
python-dateutil==2.8.1
Pillow==8.0.1
bcdata==0.4.3
whitenoise==5.2.0
python-keycloak==0.24.0
azure-storage-blob==12.8.0
openpyxl == 3.0.9
